+++ packages/playwright-config/helpers/constants.ts
@@ -1,0 +1,16 @@
+export const evergreenUrl = "http://localhost:9090";
+export const graphQLUrl = `${evergreenUrl}/graphql/query`;
+
+/**
+ * Defines user credentials that can be used to log in to the application.
+ * - admin: A user who has admin permissions and is a superuser.
+ * - privileged: A user who has admin permissions, but is not a superuser.
+ * - regular: A user who has no admin permissions.
+ */
+export const users = {
+  admin: { username: "admin", password: "password" },
+  privileged: { username: "privileged", password: "password" },
+  regular: { username: "regular", password: "password" },
+};
+
+export const toastDataCy = "toast";
