+++ packages/playwright-config/package.json
@@ -4,6 +4,10 @@
   "main": "playwright.config.ts",
   "type": "module",
   "private": true,
+  "exports": {
+    ".": "./playwright.config.ts",
+    "./helpers": "./helpers/index.ts"
+  },
   "dependencies": {
     "@playwright/test": "^1.49.1"
   },
